Rating_List_Size

+++ app/src/main/java/com/kodex/guide/ui/addscreen/data/Book.kt
@@ -24,7 +24,7 @@ data class Book(
     val deleteDate: Int = 0,
     val village: String = "",
     val delivery: Boolean = false,
-    val ratingsList: List<Double> = emptyList(),
+    val ratingsList: List<Int> = emptyList(),
 
     )
